+++ CHANGES.md
@@ -7,6 +7,7 @@ Features ✨:
 Improvements 🙌:
  - Add "show password" in import Megolm keys dialog
  - Visually disable call buttons in menu and prohibit calling when permissions are insufficient (#2112)
+ - Use cache for user color
 
 Bugfix 🐛:
  - Long message cannot be sent/takes infinite time & blocks other messages #1397